requirements.txt
django
coverage
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pip install django-shortcuts
